fix: update last_review date in seed script
Without a last_review date, it causes an error when we use
f.repeat to schedule for the next card.

This is because the constructor new SchedulingCard() in the
ts-fsrs library uses the last_review date for cards that are
not State.New to calculate the elapsed_days field.

By seeding a review_date in the past, this eliminates the error.

In the future, we should seed cards randomly using the ts-fsrs
library to initialise a card in the past.