fix: align grid items to top
This issue was caused by the `min-content_1fr` tailwind class in the
layout element.

This caused the second row to expand to the maximum possible size,
pushing the cards (which were on the third row), to the bottom of the
page.

To solve this, we put the title and the actions in the same row such
that the second row becomes the cards instead, and the cards
expand to fill the available space.