feat: add user roles
User roles will allow us to restrict the usage for certain features,
like OpenAI.

This is important as we don't want any random user to end up consuming
all of our OpenAI credits.

Note that certain commands in the package.json have also been updated
to work with the newest version of drizzle-kit.