feat: add database indices
Many of the existing queries required full scans of the table
for every request, which is an expensive operation and also
unnecessarily adds to our limit on Turso.

Let's add indices on the most used foreign keys, such as `card_id` and
`user_id` such that the query planner can more efficiently execute
queries.

See
https://turso.tech/blog/tips-for-maximizing-your-turso-billing-allowances-48a0fca163e9
the blog post by turso for more information.

Note: For some reason, indices on the columns used in `ORDER BY` don't
work.
When we execute these queries, we still see that sqlite is using a
TEMP B-TREE for these operations.
This might be due to the complexity of the queries - the joins and the
filtering conditions - though more investigation is required.