fix: handle form reset for create many
After clicking "create all" in the create many cards form,
the next time you click "create all", the deckIds are not submitted.

There are 2 bugs here.
1. `form.reset()` should not reset the deck IDs. It should preserve
the deck ids for convenience for adding multiple sets of cards.

2. The `MultiSelect` component should update its values to reflect the
current form's state. This means we must watch for changes in the form
values in the MultiSelect form component, and propagate those changes
to the MultiSelect.