Sort new cards by uuid
This would product a more "random"-ish ordering, which is helpful
for not showing too many new cards of the same type together.

Cards of the same type are usually created together, but we don't want
the user to remember the same new cards all the time.

This works as its a stable sort and there is no relationship between
the uuid and the timestamp / created at.